+++ CHANGES.rst
@@ -4,6 +4,7 @@ Upcoming Version
 Removed
 -------
 - Remove deprecated Adapter and Instrument parameters :code:`preprocess_reply`, :code:`query_delay`
+- Remove deprecated Adapter methods :code:`ask`, :code:`values`, and :code:`binary_values`
 
 
 Version 0.15.0 (2025-01-15)